refactor(sql/tables/request_helpers): sql/tables/request_helpers is now sql/tables/helpers
