+++ sql/tables/helpers/job_request.py
